Switch to shifted real to complex FFT for ~2x speedup
I also implemented a conventional complex FFT (rfft), but it's
annoying since the state is nonuniform (most values complex, 2 values
real).  Instead, we use the shifted real to complex FFT

  y(k) = sum_j e^(2𝜋ij(k+½)/n) x(j)

for which all values come in complex conjugate pairs.  This took a bit
of work to derive, mostly due to my clumsiness, but should be roughly
the same speed as an optimized real FFT on CPU and slightly faster on
GPU due to uniformity.  One of the butterflies is more expensive (~2.5
twiddles rather than 1), but the rest are equivalent to normal FFT
butterflies.  Happily, we get a nice 2x speedup:

  before (fft):  k 14, 36.1 s: mu = 1.786389717, error = 2.11e-14
  after (srfft): k 14, 17.9 s: mu = 1.786389717, error = 1.82e-14

though I think the nearness to 2x may be a consequence of not
precomputing twiddle factors just yet.