Track known series zeros
The single "terms" count is now replaced by "known" and "nonzero",
the number of terms whose value we know and the number of terms which we
know that might be nonzero.  The remaining known-nonzero terms are known
to be zero, and represented implicitly.

This gives very little speed and increase and an ambiguous accuracy change:

  before:
    k 15, 13.7 s: mu = 1.766837674, error = 9.77e-15
    k 16, 33.4 s: mu = 1.753375772, error = 4.29e-14
  after:
    k 15, 13.1 s: mu = 1.766837674, error = 2.82e-14
    k 16, 32.1 s: mu = 1.753375772, error = 3.55e-15

I'm going to keep it because it seems like it should be a net win, even
if it's small, and conceptually it's nice.